requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
djangorestframework==3.11.0
django-cors-headers==3.2.1
drf-yasg==1.17.1
idna==2.9
numpy==1.18.1
pandas==1.0.0
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
